% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_exprs}
\alias{add_exprs}
\title{Partially bold a string}
\usage{
add_exprs(...)
}
\arguments{
\item{...}{input argument for list function}
}
\value{
Expression
}
\description{
Partially bold a string
}
\examples{
add_exprs("test_bold)", "not bold")
}
