% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{br_charts_theme}
\alias{br_charts_theme}
\title{BR charts theme}
\usage{
br_charts_theme(
  base_family = "",
  base_font_size = 9,
  base_stroke = 1,
  margin = 1,
  get_fonts = control_fonts,
  get_colors = colfun()[["control_palettes"]],
  axis_text_x = ggplot2::element_text(colour = black),
  axis_line = ggplot2::element_line(colour = black, size = stroke_size),
  axis_title_y = ggplot2::element_text(),
  axis_text_y_left = ggplot2::element_text(margin = ggplot2::margin(t = 0, r = spacing/2,
    l = 0, b = 0, unit = "pt")),
  legend_position = "top",
  panel_grid_minor = ggplot2::element_line(colour = grey_2, size = stroke_size/2,
    linetype = "dashed"),
  panel_grid_major = ggplot2::element_line(colour = grey_2, size = stroke_size/2,
    linetype = "dashed"),
  ...
)
}
\arguments{
\item{base_family}{\itemize{
\item font
}}

\item{base_font_size}{(unit)\cr Font-size of normal paragraph text.}

\item{base_stroke}{(unit)\cr line thickness}

\item{margin}{(unit)\cr margin around entire plot (unit with the sizes of the
top, right, bottom, and left margins)}

\item{get_fonts}{fonts}

\item{get_colors}{colors}

\item{axis_text_x}{tick labels along axes}

\item{axis_line}{all line elements}

\item{axis_title_y}{labels of axes}

\item{axis_text_y_left}{tick labels along axes}

\item{legend_position}{position}

\item{panel_grid_minor}{grid lines}

\item{panel_grid_major}{grid lines}

\item{...}{Additional arguments passed to other methods.}
}
\value{
theme for chart
}
\description{
BR charts theme
}
