% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_calculation.R
\name{calculate_diff_bin}
\alias{calculate_diff_bin}
\title{CI for absolute risk for binary outcomes}
\usage{
calculate_diff_bin(prop1, prop2, n1, n2, cl = 0.95)
}
\arguments{
\item{prop1}{(\code{numeric})\cr Proportion of cases in active treatment}

\item{prop2}{(\code{numeric})\cr Proportion of cases in comparator treatment}

\item{n1}{(\code{numeric})\cr Total number of subjects in active treatment}

\item{n2}{(\code{numeric})\cr Total number of subjects in comparator treatment}

\item{cl}{(\code{numeric})\cr confidence level}
}
\description{
Derive mean difference and associated confidence intervals
for binary outcomes
}
\examples{
calculate_diff_bin(
  prop1 = .45, prop2 = 0.25, n1 = 500, n2 = 500,
  cl = 0.95
)
}
