% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_calculation.R
\name{calculate_diff_con}
\alias{calculate_diff_con}
\title{CI for treatment difference in continuous outcomes}
\usage{
calculate_diff_con(mean1, mean2, sd1, sd2, n1, n2, cl = 0.95)
}
\arguments{
\item{mean1}{(\code{numeric})\cr Mean of measure in active treatment}

\item{mean2}{(\code{numeric})\cr Mean of measure in comparator treatment}

\item{sd1}{(\code{numeric})\cr Standard deviation of measure in active treatment}

\item{sd2}{(\code{numeric})\cr Standard deviation of measure in comparator
treatment}

\item{n1}{(\code{numeric})\cr Total number of subjects in active treatment}

\item{n2}{(\code{numeric})\cr Total number of subjects in comparator treatment}

\item{cl}{(\code{numeric})\cr confidence level}
}
\description{
Derive mean difference and associated confidence intervals
for continuous outcomes
}
\examples{
calculate_diff_con(
  mean1 = 0.6, mean2 = 0.5, sd1 = 0.1, sd2 = 0.3,
  n1 = 400, n2 = 500, cl = 0.95
)
}
