% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feature.R
\name{check_effects_table}
\alias{check_effects_table}
\title{Check if data contains required features to run a specific plot}
\usage{
check_effects_table(df)
}
\arguments{
\item{df}{(\code{data.frame}) dataset - effect table}
}
\value{
missing features
}
\description{
Check if data contains required features to run a specific plot
}
\details{
This function verifies whether the input dataset (\code{df})
contains all the necessary features required to generate a specific plot.
It checks for the existence of necessary features as well as the types
and values of features and display log messages if any feature does not
conform to the defined rules. This ensures that the dataset is suitable for
the intended visualization.
}
\examples{
check_effects_table(brdata)
}
