% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feature.R
\name{check_feature}
\alias{check_feature}
\title{intermediate function used to display log messages
check if a specific feature exist in the data}
\usage{
check_feature(
  data,
  feature,
  plots,
  func,
  na_check,
  values,
  check_same,
  check_range,
  check_positive,
  check_unique
)
}
\arguments{
\item{data}{dataset}

\item{feature}{(\code{data.frame}) feature to display analysis}

\item{plots}{(\code{function}) type of analysis (graph)
either \code{forest}, \code{tradeoff}, \code{contour}, \code{value_tree}}

\item{func}{function to check data type}

\item{na_check}{(\code{logical}) check if the feature has missing values}

\item{values}{(\code{vector}) check if the feature contains specified values}

\item{check_same}{(\code{logical}) check if the feature has the same value across all rows}

\item{check_range}{(\code{vector}) check if the feature is in the specified range}

\item{check_positive}{(\code{logical}) check if the feature is positive}

\item{check_unique}{(\code{vector}) check if unique values of a feature is associated with unique values of linked features}
}
\value{
error message(s), if any
}
\description{
intermediate function used to display log messages
check if a specific feature exist in the data
}
\details{
DETAILS
}
