% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comp_outcome}
\alias{comp_outcome}
\title{Example composite outcome data used for Figure 12}
\format{
A data frame with 1800 rows and 6 variables
\describe{
\item{usubjid}{Subject ID}
\item{visit}{Visit}
\item{trtn}{Treatment arms in numeric type}
\item{trt}{Treatment arms in character type}
\item{brcatn}{Category of composite outcome in numeric type}
\item{brcat}{Category of composite outcome in character type}

}
}
\usage{
comp_outcome
}
\description{
Example composite outcome data used for Figure 12
}
\keyword{datasets}
