% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{control_fonts}
\alias{control_fonts}
\title{Control Fonts}
\usage{
control_fonts(base_font_size = 9, h1 = 12, h2 = 10, label = base_font_size + 1)
}
\arguments{
\item{base_font_size}{(unit)\cr Font-size of normal paragraph text.

By default is 9pt.

By editing this value, all the other parameters get updated too.}

\item{h1}{(unit)\cr Font-size of title of the graph,
dependent of \code{base_font_size}. By default, 12pt.}

\item{h2}{(unit)\cr Font-size of subtitle of the graph,
dependent of \code{base_font_size}. By default, 10pt.}

\item{label}{(unit)\cr Font-size of text that's called outside the ggplot
theme. It should render at the same font size as base_font_size}
}
\description{
\code{control_fonts()} calculates font-sizes depending on output size
}
\details{
It returns a list with the font sizes of:
\itemize{
\item \strong{p} - normal text elements within the ggplot theme.
\item \strong{h1} - titles.
\item \strong{h2} - subtitles.
\item \strong{label} - for text elements outside the ggplot theme. If you want to add
an annotation or a label to your plot, use the font size \code{label}. By default,
it renders at 10pt.
\item \strong{rel} - for legend elements.
}
}
\examples{
control_fonts(base_font_size = 10)
}
