% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cumexcess}
\alias{cumexcess}
\title{Example cumulative excess plot data used for Figure 13}
\format{
A data frame with 880 rows and 9 columns:
\describe{
\item{eventtime}{Simulated event times}
\item{diff}{Simulated difference in active/control effects}
\item{obsv_duration}{Duration of observational period}
\item{obsv_unit}{Unit length of observational period}
\item{outcome}{Specifies Benefit/Risk}
\item{eff_diff_lbl}{Label for effect difference}
\item{n}{Number of subjects}
\item{effect}{Specifies active/control effect}
\item{eff_code}{0/1 depicting control/active effects}

}
}
\usage{
cumexcess
}
\description{
Example cumulative excess plot data used for Figure 13
}
\keyword{datasets}
