% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_barchart.R
\name{divergent_stacked_barchart}
\alias{divergent_stacked_barchart}
\title{Divergent Stacked Bar Chart}
\usage{
divergent_stacked_barchart(
  data,
  chartcolors,
  favcat,
  unfavcat,
  ylabel = "Visit",
  base_font_size = 9
)
}
\arguments{
\item{data}{\code{dataframe} a data frame with a minimum of 4 variables named the
following:
\enumerate{
\item usubjid: unique subject ID
\item visit: visit ID
\item trt: treatment group
\item brcat: composite benefit-risk category
}}

\item{chartcolors}{\code{vector} a vector of colors, the same number of levels as
the brcat variable}

\item{favcat}{\code{vector} a vector of favorable categories in the desired plot
order}

\item{unfavcat}{\code{vector} a vector of unfavorable categories in the desired
plot order}

\item{ylabel}{\code{character} y label name, default is "Visit"}

\item{base_font_size}{Numeric; base font size in points for all text
elements in the plot (default: 9).}
}
\value{
a ggplot object
}
\description{
Divergent Stacked Bar Chart
}
\examples{
divergent_stacked_barchart(
  data = comp_outcome,
  chartcolors = colfun()$fig12_colors,
  favcat = c("Benefit larger than threshold, with AE",
  "Benefit larger than threshold, w/o AE"),
  unfavcat = c("Withdrew",
  "Benefit less than threshold, w/o AE",
  "Benefit less than threshold, with AE"),
  ylabel = "Study Week"
)

}
