% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{effects_table}
\alias{effects_table}
\title{Example treatment effect table}
\format{
A data frame with ...
}
\usage{
data(effects_table)
}
\description{
Sample data frame for use in plot examples
}
