% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancement_cumexcess.R
\name{gensurv}
\alias{gensurv}
\title{Simulate data (utilized for function tests)}
\usage{
gensurv(seed, n1, n2, obsv_duration, lambda1, lambda2, unit = "Months")
}
\arguments{
\item{seed}{A numerical input or object that will set the standard for data
simulation.}

\item{n1}{A numerical input or object that represents the desired total
population for the simulated data (for treatment group 1).}

\item{n2}{A numerical input or object that represents the desired total
population for the simulated data (for treatment group 2).}

\item{obsv_duration}{A numerical input or object that is the maximum time
the observational period will span for the simulated data.}

\item{lambda1}{A numerical input or object that is the first lambda that will
be used in the data simulation (used as the first parameter for simulation).}

\item{lambda2}{A numerical input or object that is the second lambda that
will be used in the data simulation (used as the second parameter
for simulation).}

\item{unit}{A non-numerical input that specifies the length of the
observational period ("Week", "Month", etc.).}
}
\value{
A simulated dataframe.
}
\description{
Simulate data (utilized for function tests)
}
\examples{
gensurv(111, 2000, 1000, 36, .005, .0048, "Weeks")
}
