% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancement_cumexcess.R
\name{gensurv_table}
\alias{gensurv_table}
\title{Create a table that corresponds to the cumulative excess plot}
\usage{
gensurv_table(
  df_table,
  base_subjects,
  visits,
  fig_colors = c("#0571b0", "#ca0020")
)
}
\arguments{
\item{df_table}{A dataframe with 6 variables named the following:
\enumerate{
\item obsv_duration: A variable that specifies the duration of the
observational period (numerical).
\item n: A vector containing a number of subjects who experienced an event
at a given time (numerical).
\item effect: specifies between an active or control effect.
\item outcome: specifies whether the an outcome should be classified as a
"Benefit" or "Risk" (this must have either "Benefit" or "Risk" as values).
\item eff_code: 0 for control and 1 for active effect.
\item subjects: A vector containing the total number of active/placebo subjects
in the study at a given time.
}}

\item{base_subjects}{A numerical input that specifies the baseline proportion
of subjects in the study (for example, "per 100 subjects")}

\item{visits}{A numerical input that is the length between observational
periods.}

\item{fig_colors}{Allows the user to change the colors of the table
(defaults are provided). Must be vector of length 2, with color corresponding
to benefit second and risk first.}
}
\value{
A table.
}
\description{
Create a table that corresponds to the cumulative excess plot
}
\examples{
gensurv_table(cumexcess, 100, 6)
}
