% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggsave_custom}
\alias{ggsave_custom}
\title{Wrapper to ggsave: Save a ggplot (or other grid object) with sensible
defaults}
\usage{
ggsave_custom(
  save_name,
  inplot = NULL,
  imgpath = ".",
  wdth = 7,
  hght = 4.1,
  unts = "in",
  bgcol = "white",
  dpi = 600,
  web_suffix = FALSE,
  scale_fonts = FALSE,
  base_font_size = NULL,
  ...
)
}
\arguments{
\item{save_name}{File name to create on disk.}

\item{inplot}{Plot to save, defaults to last plot displayed.}

\item{imgpath}{Path of the directory to save plot to: path}

\item{wdth}{width of plot}

\item{hght}{height of plot}

\item{unts}{units of plot}

\item{bgcol}{Background color. If NULL, uses the plot.background fill value
from the plot theme.}

\item{dpi}{Resolution in dots per inch (default: 600).}

\item{web_suffix}{If TRUE, also saves a low-res version with "_web" suffix
(default: FALSE).}

\item{scale_fonts}{Deprecated parameter. Font scaling is now handled
by passing base_font_size directly to plotting functions.}

\item{base_font_size}{Deprecated parameter. Font scaling is now handled
by passing base_font_size directly to plotting functions.}

\item{...}{Other arguments passed on to the graphics device function,
as specified by device.}
}
\description{
Adds customized defaults to ggsave for the BRAP Journal requirements
}
