% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_br_calculated_ci}
\alias{prepare_br_calculated_ci}
\title{Prepare data analysis for binary and continuous outcomes with Calculated
interval confidence
identifies whether the dataframe is for Benefit or Risk analysis}
\usage{
prepare_br_calculated_ci(df, colname1, colname2, cl = 0.95, func)
}
\arguments{
\item{df}{(\code{data.frame}) dataset
either \code{df_benefit} (selected benefit)
or \code{df_risk} (select risk).}

\item{colname1}{(\code{character}) feature to fetch for the analysis
either \code{Mean}, \code{Prop}, \code{Rate}}

\item{colname2}{(\code{character}) feature to fetch for the analysis
either \code{nPat}, \code{Py}}

\item{cl}{(\code{numeric}) confidence level}

\item{func}{(\code{function}) function used to calculate metrics (or BR points)}
}
\value{
data frame for specified type of analysis
}
\description{
Prepare data analysis for binary and continuous outcomes with Calculated
interval confidence
identifies whether the dataframe is for Benefit or Risk analysis
}
\details{
DETAILS
}
