% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_br_supplied_ci}
\alias{prepare_br_supplied_ci}
\title{Prepare data analysis for binary and continuous outcomes with Supplied
interval confidence
identifies whether the dataframe is for Benefit or Risk analysis}
\usage{
prepare_br_supplied_ci(df, colname, metric_name, func)
}
\arguments{
\item{df}{(\code{data.frame}) dataset
either \code{df_benefit} (selected benefit)
or \code{df_risk} (select risk).}

\item{colname}{(\code{character}) feature to fetch for the analysis
either \code{Mean}, \code{Prop}, \code{Rate}}

\item{metric_name}{(\code{character}) metric for which we must fetch the
confidence interval if supplied (taken from the effect table)
either \code{Diff}, \code{RelRisk}, \code{OddsRatio}, \code{Diff_Rates}}

\item{func}{(\code{function}) function used to calculate metrics (or BR points)}
}
\value{
data frame for specified type of analysis
}
\description{
Prepare data analysis for binary and continuous outcomes with Supplied
interval confidence
identifies whether the dataframe is for Benefit or Risk analysis
}
\details{
DETAILS
}
