% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_plot.R
\name{prepare_tradeoff_data}
\alias{prepare_tradeoff_data}
\title{Prepare data for the tradeoff plot}
\usage{
prepare_tradeoff_data(
  data,
  filter,
  category,
  benefit,
  risk,
  ci_method,
  cl,
  type_risk,
  type_graph
)
}
\arguments{
\item{data}{(\code{data.frame}) dataset}

\item{filter}{(\code{character}) selected filter}

\item{category}{(\code{character}) selected category}

\item{benefit}{(\code{character}) selected benefit outcome}

\item{risk}{(\code{character}) selected risk outcome}

\item{ci_method}{(\code{character}) selected method to display
confidence intervals}

\item{cl}{(\code{numeric}) confidence level}

\item{type_risk}{(\code{character}) selected way to display risk outcomes
(crude proportions, Exposure-adjusted rates (per 100 PYs))}

\item{type_graph}{(\code{character}) selected way to display binary outcomes
(Absolute risk, Relative risk, Odds ratio)}
}
\value{
df_br (\code{data.frame}) benefit/risk metrics for all treatment given
the selected benefit (respectively risk) outcome
}
\description{
Prepare data for the tradeoff plot
}
\details{
DETAILS
}
