% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeoff_plot.R
\name{prepare_tradeoff_plot}
\alias{prepare_tradeoff_plot}
\title{Prepare trade-off plot}
\usage{
prepare_tradeoff_plot(
  myplot,
  data,
  df_br,
  drug_status,
  filter,
  ci,
  chartcolors
)
}
\arguments{
\item{myplot}{raw plot}

\item{data}{(\code{data.frame}) input dataset}

\item{df_br}{(\code{data.frame}) processed dataset}

\item{drug_status}{(\code{character}) selected status of drug to display
(Approved, Test)}

\item{filter}{(\code{character}) selected filter}

\item{ci}{(\code{character}) selected choice to display confidence intervals
or not (Yes, No)}

\item{chartcolors}{(\code{vector}) a vector of colors, the same number of levels
as the number of treatments}
}
\description{
Add points and dual CIs to trade-off plot
}
