% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{publication_geom_text_size}
\alias{publication_geom_text_size}
\title{Convert points to ggplot geom text size units}
\usage{
publication_geom_text_size(point_size)
}
\arguments{
\item{point_size}{Font size in points.}
}
\value{
Numeric size value for \code{geom_text(size = ...)}.
}
\description{
Convert points to ggplot geom text size units
}
