% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relmax}
\alias{relmax}
<<<<<<< HEAD
\title{Derive maximum boundary value for axis
Derive boundary value to include all values}
=======
\title{Derive maximum boundary value for axis to include all values}
>>>>>>> main
\usage{
relmax(rmax, type_scale)
}
\arguments{
\item{rmax}{(\code{numeric}) number to evaluate}

\item{type_scale}{(\code{character}) selected scale display type}
}
\value{
numeric
}
\description{
<<<<<<< HEAD
Derive maximum boundary value for axis
Derive boundary value to include all values
=======
Derive maximum boundary value for axis to include all values
>>>>>>> main
}
\examples{
relmax(0.5, "Free")
relmax(0.5, "Fixed")
relmax(-0.3, "Free")
relmax(-0.3, "Fixed")
}
