% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Create a scatterplot from a given dataframe.}
\usage{
scatter_plot(
  df_diff,
  outcome,
  MAB,
  MAR,
  type = "histogram",
  fig_colors = colfun()$fig11_colors
)
}
\arguments{
\item{df_diff}{A dataframe containing two vectors, each of which displays the
difference between incremental probabilities in active and control effects
for a specified outcome.}

\item{outcome}{A vector of two strings that describes the two outcomes
associated with the difference in active and control effects, where the first
outcome corresponds to \code{diff1} and the second to \code{diff2}.}

\item{MAB}{A numerical value that specifies the mimimum acceptable benefit.}

\item{MAR}{A numerical value that specifies the maximum acceptable risk.}

\item{type}{Type of marginal plot to show. One of: density, histogram,
boxplot, violin, densigram (a 'densigram' is when a density plot is overlaid
on a histogram). By default, histogram is displayed.}

\item{fig_colors}{Allows user to change colors of the figure (defaults are
provided). Must be a vector of length 3, with the first color corresponding
to the scatter plot points, the second corresponding to the overall mean, and
third to the written probability text color.}
}
\value{
A scatterplot.
}
\description{
Create a scatterplot from a given dataframe.
}
\examples{
outcome <- c("Benefit", "Risk")
scatter_plot(scatterplot, outcome, MAB = 0.2, MAR = 0.6)

}
