% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scatterplot}
\alias{scatterplot}
\title{Example scatterplot data used for Figure 11}
\format{
A data frame with 500 rows and 2 columns:
\describe{
\item{bdiff}{Simulated difference in incremental probabilities for
active/control effects and outcome "Benefit"}
\item{rdiff}{Simulated difference in incremental probabilities for
active/control effects and outcome "Risk"}

}
}
\usage{
scatterplot
}
\description{
Example scatterplot data used for Figure 11
}
\keyword{datasets}
