% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_barchart.R
\name{stacked_barchart}
\alias{stacked_barchart}
\title{Stacked Bar Chart}
\usage{
stacked_barchart(data, chartcolors, ylabel = "Visit", base_font_size = 9)
}
\arguments{
\item{data}{\code{dataframe} a data frame with a minimum of 4 variables named the
following:
\enumerate{
\item usubjid: unique subject ID
\item visit: visit ID
\item trt: treatment group
\item brcat: composite benefit-risk category
}}

\item{chartcolors}{\code{vector} a vector of colors, the same number of levels as
the brcat variable}

\item{ylabel}{\code{character} y label name, default is "Visit"}

\item{base_font_size}{Numeric; base font size in points for all text
elements in the plot (default: 9).}
}
\value{
a ggplot object
}
\description{
Stacked Bar Chart
}
\examples{
stacked_barchart(
  data = comp_outcome,
  chartcolors = colfun()$fig12_colors,
  ylabel = "Study Week"
)

}
