% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feature.R
\name{check_effects_table}
\alias{check_effects_table}
\title{Check if data contains required features to run a specific plot}
\usage{
check_effects_table(df)
}
\arguments{
\item{df}{(\code{data.frame}) dataset - effect table}
}
\value{
missing features
}
\description{
Check if data contains required features to run a specific plot
}
\details{
DETAILS
}
\examples{
check_effects_table(brdata)
}
