% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{labs_bold}
\alias{labs_bold}
\title{Create expression}
\usage{
labs_bold(cond, bold, nonbold)
}
\arguments{
\item{cond}{(\code{numeric}) expected conditional variable bold(1), not to bold(0)}

\item{bold}{(\code{character})  level to bold}

\item{nonbold}{(\code{character})\cr which level to bold.}
}
\description{
Selectively bold label for \code{\link[ggplot2:ggplot2-package]{ggplot2::ggplot2()}}.
}
\details{
The function bold text in variable (\code{bold}) and concatenates it
with string in (\code{nonbold}) and returns a \code{dataframe}.
}
\examples{
library(dplyr)
library(ggplot2)

xxx <- tribble(
  ~x, ~z, ~w, ~y,
  1, "BOLD_AA", " plain", 1,
  2, "b", "b", 0,
  3, "c", "c", 0
)
ggplot(xxx, aes_string(x = "x", y = "z")) +
  geom_point() +
  scale_y_discrete(
    label = labs_bold(cond = xxx[["y"]], xxx[["z"]], nonbold = xxx[["w"]])
  )

}
\seealso{
\code{?plotmath}.
}
