% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relmin}
\alias{relmin}
\title{Derive minimum boundary value for axis
Derive boundary value to include all values}
\usage{
relmin(rmin, type_scale)
}
\arguments{
\item{rmin}{(\code{numeric}) number to evaluate}

\item{type_scale}{(\code{character}) selected scale display type}
}
\value{
numeric
}
\description{
Derive minimum boundary value for axis
Derive boundary value to include all values
}
\examples{
relmin(0.5, "Free")
relmin(0.5, "Fixed")
relmin(-0.3, "Free")
relmin(-0.3, "Fixed")
}
